Add README with Table of Contents
Introduce an enhanced README file that includes a table of contents for easy navigation. The README provides an in-depth overview of the Weather App, highlighting its features, functionality & a live demo link .

The addition of the table of contents enhances the reader's experience by offering structured access to different sections of the README. This contribution aims to facilitate clear understanding and efficient exploration of the Weather App's capabilities and usage.